add createBillForm and BillsList

+++ packages/client/src/components/Account/AccountCard.tsx
@@ -12,11 +12,13 @@ import DeleteIcon from "@mui/icons-material/Delete";
 import { deleteAccount } from "../../api/account";
 import { FindAccount } from "../../../../contracts";
 import {
+  $accountsStore,
   $pickedAccount,
   setPickedAccount,
   updateAccountsStoreFx,
 } from "../../store/AccountStore";
-
+import { $usersBillsStore } from "../../store/BillStore";
+import { addError } from "../../store/ErrorsStore";
 
 type AccountCardProps = {
   account: FindAccount.Response;
@@ -27,9 +29,21 @@ export default function AccountCard(props: AccountCardProps) {
   const currency = useStore($currencyStore).find(
     (cur) => cur.id === props.account.currencyId
   );
+  const billsList = useStore($usersBillsStore);
   const handleDeleteAccount = async () => {
-    await deleteAccount(props.account.id);
-    updateAccounts(null);
+    if (
+      billsList.find((bill) => bill.accountId === props.account.id) ===
+      undefined
+    ) {
+      await deleteAccount(props.account.id);
+      updateAccounts(null);
+    } else {
+      const error = new Error();
+      error.message = "You cant delete this account beacuse it have bills";
+      error.name = "Delete account error";
+      addError(error);
+      console.log(error.message);
+    }
   };
   const isPicked =
     useStore($pickedAccount)?.id === props.account.id ? styles.picked : "";
